move UserTransactionType.kt to helper package
