+++ wallet/src/main/kotlin/net/thechance/wallet/api/dto/transaction/StatementData.kt
@@ -1,5 +1,6 @@
 package net.thechance.wallet.api.dto.transaction
 
+import net.thechance.wallet.service.helper.UserTransactionType
 import java.time.LocalDateTime
 import java.util.*
