+++ wallet/src/main/kotlin/net/thechance/wallet/service/helper/TransactionFilterParams.kt
@@ -1,6 +1,5 @@
 package net.thechance.wallet.service.helper
 
-import net.thechance.wallet.api.dto.transaction.UserTransactionType
 import net.thechance.wallet.entity.Transaction
 import java.time.LocalDate
